Move switchdb command to owner cog
